Move code out of header
